DataRepository interface, services renames
