+++ lib/logic/active_user/active_user_cubit.dart
@@ -6,8 +6,8 @@ import 'package:mongo_dart/mongo_dart.dart';
 import 'package:fokus/model/db/user/user.dart';
 import 'package:fokus/model/db/user/user_role.dart';
 import 'package:fokus/model/ui/ui_user.dart';
-import 'package:fokus/services/database/data_repository.dart';
-import 'package:fokus/services/settings/app_config_repository.dart';
+import 'package:fokus/services/data/data_repository.dart';
+import 'package:fokus/services/app_config/app_config_repository.dart';
 
 part 'active_user_state.dart';
