+++ lib/logic/app_init/app_init_cubit.dart
@@ -1,11 +1,12 @@
 import 'package:cubit/cubit.dart';
-import 'package:fokus/services/plan_repeatability_service.dart';
 import 'package:get_it/get_it.dart';
 
-import 'package:fokus/services/database/data_repository.dart';
+import 'package:fokus/services/data/db_data_repository.dart';
 import 'package:fokus/services/remote_config_provider.dart';
-import 'package:fokus/services/settings/app_config_repository.dart';
-import 'package:fokus/services/settings/app_shared_preferences_provider.dart';
+import 'package:fokus/services/app_config/app_config_repository.dart';
+import 'package:fokus/services/app_config/app_shared_preferences_provider.dart';
+import 'package:fokus/services/data/data_repository.dart';
+import 'package:fokus/services/plan_repeatability_service.dart';
 
 import 'app_init_state.dart';
 
@@ -15,7 +16,7 @@ class AppInitCubit extends Cubit<AppInitState> {
 	AppInitCubit() : super(AppInitInProgress()) {
 		_provider.registerSingleton<RemoteConfigProvider>(RemoteConfigProvider());
 		_provider.registerSingleton<AppConfigRepository>(AppConfigRepository(AppSharedPreferencesProvider())..initialize());
-		_provider.registerSingleton<DataRepository>(DataRepository());
+		_provider.registerSingleton<DataRepository>(DbDataRepository());
 		_provider.registerSingleton<PlanRepeatabilityService>(PlanRepeatabilityService());
 		initializeApp();
 	}
@@ -24,7 +25,7 @@ class AppInitCubit extends Cubit<AppInitState> {
 		// TODO Differentiate between no internet connection and db access error
 
 		await _provider<RemoteConfigProvider>().initialize().then(
-			(_) => _provider<DataRepository>().initialize(_provider<RemoteConfigProvider>().dbAccessString)
+			(_) => _provider<DataRepository>().initialize()
 		).then((_) => emit(AppInitSuccess())).catchError((error) => emit(AppInitFailure(error)));
 	}
 }
