+++ lib/logic/caregiver_panel/caregiver_panel_cubit.dart
@@ -1,14 +1,14 @@
 import 'package:cubit/cubit.dart';
 import 'package:equatable/equatable.dart';
-import 'package:fokus/model/ui/ui_caregiver.dart';
 import 'package:get_it/get_it.dart';
+import 'package:mongo_dart/mongo_dart.dart';
 
 import 'package:fokus/logic/active_user/active_user_cubit.dart';
 import 'package:fokus/services/plan_repeatability_service.dart';
-import 'package:fokus/services/database/data_repository.dart';
+import 'package:fokus/services/data/data_repository.dart';
 import 'package:fokus/model/ui/ui_child.dart';
 import 'package:fokus/model/db/date/date.dart';
-import 'package:mongo_dart/mongo_dart.dart';
+import 'package:fokus/model/ui/ui_caregiver.dart';
 
 part 'caregiver_panel_state.dart';
