+++ lib/services/data/mongodb_provider.dart
@@ -2,7 +2,7 @@ import 'package:fokus/model/db/collection.dart';
 
 import 'package:mongo_dart/mongo_dart.dart';
 
-class MongoClient {
+class MongoDbProvider {
 	var _useLocalhost = false;
 	var _localhostConfig = 'mongodb://root:pass@192.168.0.10:27017/admin';
 
@@ -13,13 +13,8 @@ class MongoClient {
 		return _db.open(secure: !_useLocalhost);
 	}
 
-	Future<Map<String, dynamic>> queryOne(Collection collection, [SelectorBuilder selector]) {
-		return _db.collection(collection.name).findOne(selector);
-	}
-
-	Future<Stream<Map<String, dynamic>>> query(Collection collection, [SelectorBuilder selector]) async {
-		return _db.collection(collection.name).find(selector);
-	}
+	Future<Map<String, dynamic>> queryOne(Collection collection, [SelectorBuilder selector]) => _db.collection(collection.name).findOne(selector);
+	Future<Stream<Map<String, dynamic>>> query(Collection collection, [SelectorBuilder selector]) async => _db.collection(collection.name).find(selector);
 
 	Future<int> count(Collection collection, [SelectorBuilder selector]) async => _db.collection(collection.name).count(selector);
 	Future<bool> exists(Collection collection, [SelectorBuilder selector]) async => await count(collection, selector) > 0;
